requirements.txt
beautifulsoup4==4.14.2
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
filelock==3.20.0
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
ruff==0.14.2
setuptools==80.9.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
soupsieve==2.8
tweepy==4.16.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
